+++ package.json
@@ -46,7 +46,7 @@
     "@angular/router": "9.0.0",
     "@ng-bootstrap/ng-bootstrap": "6.0.0",
     "ajv": "^5.2.2",
-    "amo-tools-suite": "0.7.0",
+    "amo-tools-suite": "../AMO-Tools-Suite",
     "bootstrap": "4.3.1",
     "c3": "^0.4.18",
     "core-js": "2.5.7",
