+++ src/app/motor-inventory/motor-inventory-summary/summary-filter/summary-filter.component.css
@@ -1,0 +1,26 @@
+.navbar{
+    background: white;
+    border-bottom: solid 4px;
+    font-size: small;
+}
+.list-group-item{
+    padding: 8px;
+    font-size: small;
+}
+
+.list-group-item.active{
+    color: white;
+    background: #2E4053;
+}
+
+a.list-group-item:hover, .nav-link{
+    cursor: pointer;
+}
+
+a.list-group-item.active:hover{
+    color: white;
+}
+
+.nav-link.dropdown-toggle:hover, .nav-btn:hover{
+    background-color: #808B96;
+}
